+++ R/prepareCruiseSeriesList.R
@@ -7,10 +7,13 @@
 prepareCruiseSeriesList <- function() {
   
   pb <- utils::txtProgressBar(max = 32, style = 3)
+  on.exit(close(pb), add = TRUE)
   utils::setTxtProgressBar(pb, 1)
   
   # Read cruise reference
-  doc <- xml2::read_xml("https://reference-api.hi.no/apis/nmdapi/reference/v2/model/cruiseseries?version=2.0")
+  doc <- .reference_api_get_xml(paste0(
+    .REFERENCE_API_BASE, "/model/cruiseseries?version=2.0"
+  ))
   first <- lapply(xml2::xml_find_all(doc, "//d1:row"), function(x) {
     ch <- xml2::xml_children(x)
     y <- xml2::xml_text(ch)
@@ -29,13 +32,19 @@ prepareCruiseSeriesList <- function() {
     x <- first[[i]]
     
     seriesCode <- x[["code"]]
-    subdoc <- xml2::read_xml(paste0("https://reference-api.hi.no/apis/nmdapi/reference/v2//model/cruiseseries/", seriesCode, "/samples?version=2.0"))
+    subdoc <- .reference_api_get_xml(paste0(
+      .REFERENCE_API_BASE, "/model/cruiseseries/", seriesCode,
+      "/samples?version=2.0"
+    ))
     years <- xml2::xml_text(xml2::xml_find_all(subdoc, "//d1:sampleTime"))
     
     # Get per-year details
     cruises <- lapply(years, function(y) {
       # message(y)
-      subsubdoc <- xml2::read_xml(paste0("https://reference-api.hi.no/apis/nmdapi/reference/v2//model/cruiseseries/", seriesCode, "/samples/", y, "/cruises?version=2.0"))
+      subsubdoc <- .reference_api_get_xml(paste0(
+        .REFERENCE_API_BASE, "/model/cruiseseries/", seriesCode,
+        "/samples/", y, "/cruises?version=2.0"
+      ))
       third <- lapply(xml2::xml_find_all(subsubdoc, "//d1:row"), function(z) {ch <- xml2::xml_children(z); zz <- xml2::xml_text(ch); zzz <- xml2::xml_name(ch); names(zz) <- zzz; return(as.list(zz))})
       subret <- data.table::rbindlist(third, fill = TRUE)
       if(nrow(subret) > 0) subret[, `:=`(code = NULL, year = y, cruiseseriescode = seriesCode)]